Fixing 10um download in test, added conftest file, argumen can be passed to check_atlas_install
