+++ tests/tests/test_integration/test_registration.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 import sys
 import pytest
 
@@ -21,50 +22,12 @@ y_pix = "40"
 z_pix = "50"
 
 
-TEST_ATLAS = "allen_2017_100um"
 
 
-def download_atlas(directory):
-    download_args = [
-        "amap_download",
-        "--atlas",
-        TEST_ATLAS,
-        "--install-path",
-        directory,
-        "--no-amend-config",
-    ]
-    sys.argv = download_args
-    amap_download()
-    return directory
+#@pytest.mark.slow
+def test_register(tmpdir, test_config_path):
 
-
-def generate_test_config(atlas_dir):
-    config = os.path.join(os.getcwd(), "tests", "data", "config", "test.conf")
-    config_obj = get_config_obj(config)
-    atlas_conf = config_obj["atlas"]
-    orig_base_directory = atlas_conf["base_folder"]
-
-    with open(config, "r") as in_conf:
-        data = in_conf.readlines()
-    for i, line in enumerate(data):
-        data[i] = line.replace(
-            f"base_folder = '{orig_base_directory}",
-            f"base_folder = '{os.path.join(atlas_dir, TEST_ATLAS)}",
-        )
-    test_config = os.path.join(atlas_dir, "config.conf")
-    with open(test_config, "w") as out_conf:
-        out_conf.writelines(data)
-
-    return test_config
-
-
-@pytest.mark.slow
-def test_register(tmpdir):
-    atlas_directory = str(tmpdir)
-    download_atlas(atlas_directory)
-    test_config = generate_test_config(atlas_directory)
-
-    output_directory = os.path.join(atlas_directory, "output")
+    output_directory = os.path.join(str(tmpdir), "output")
     amap_args = [
         "amap",
         data_dir,
@@ -78,7 +41,7 @@ def test_register(tmpdir):
         "--n-free-cpus",
         "0",
         "--registration-config",
-        test_config,
+        test_config_path,
         "-d",
         data_dir,
     ]
